Create summary statistics for implicit sample

This provides a simple table of summary statistics for the implicit sample defined by Aronow and Samii (2015) doi:10.1111/ajps.12185 .

# S3 method for class 'regweight'
summary(object, df, output = "tibble", ...)

Arguments

object

Weighting model object

df

dataframe with one column for each covariate to include in the resulting table of summary statistics.

output

Desired output type. Default is to return a tibble, but can also select from "latex" and "html" to return a formatted table for inclusion in a paper or report.

...

unused

Value

One of three outputs depending on the requested type:

tibble: Returns a tibble object (see tibble::tibble()).

latex: Returns a knit_asis object (see knitr::asis_output()).

html: Returns an html object (see htmltools::HTML()).